Modified Quadratic Shepard tool: remove abs/dabs #defines (from fortran code) and replace dabs by fabs calls; these defines clash with cstdlib in case we want to use std functions somewhere in the tool library; also fix warning (size_t to int)

+++ saga-gis/src/tools/grid/grid_gridding/Interpolation_Shepard.cpp
@@ -134,7 +134,7 @@ bool CInterpolation_Shepard::On_Initialize(void)
 		m_Points[0].Get_Data(),
 		m_Points[1].Get_Data(),
 		m_Points[2].Get_Data(),
-		m_Points[0].Get_Size() - 1,
+		(int)m_Points[0].Get_Size() - 1,
 		Parameters("QUADRATIC_NEIGHBORS")->asInt(),
 		Parameters("WEIGHTING_NEIGHBORS")->asInt()
 	);
